backend/requirements.txt
fastapi
uvicorn
opencv-python
pptx
torch
torchvision
tqdm
pillow
numpy